Extract `FakeArtisanRequest` out so that we can reuse it elsewhere.
